+++ js/filters.js
@@ -1,0 +1,1 @@
+// Модуль работы с фильтрами
