+++ js/mark.js
@@ -1,0 +1,1 @@
+// Модуль работы с меткой объявления пользователя
