+++ js/similar-marks.js
@@ -1,0 +1,1 @@
+// Модуль работы с метками похожих объявлений
